+++ src/commands/queue-members/get.ts
@@ -9,7 +9,7 @@ export class queueMembersGet extends Command {
     static description = ` Retrieve a representation of the specified Queue Member.`
 
     static flags = {
-        next: flags.boolean({ char: "n", description: "Displays the next page of output." }),
+        next: flags.boolean({ hidden: true }),
         help: flags.help({ char: "h" }),
     }
 
@@ -21,7 +21,7 @@ export class queueMembersGet extends Command {
         },
         {
             name: "callId",
-            description: "ID of the Call that the Member belongs to",
+            description: "ID of the Call that the Member belongs to.",
             required: true,
         },
     ]
@@ -50,25 +50,9 @@ export class queueMembersGet extends Command {
                 throw new Errors.UndefinedResponseError()
             }
         }
-        const nextResponse = (response: FreeClimbResponse) => {
-            if (response.data) {
-                out.out(JSON.stringify(response.data, null, 2))
-            } else {
-                throw new Errors.UndefinedResponseError()
-            }
-            if (out.next === null) {
-                out.out("== You are on the last page of output. ==")
-            }
-        }
-
         if (flags.next) {
-            if (out.next === undefined || out.next === "freeclimbUnnamedTest") {
-                const error = new Errors.NoNextPage()
-                this.error(error.message, { exit: error.code })
-            } else {
-                await fcApi.apiCall("GET", { params: { cursor: out.next } }, nextResponse)
-            }
-            return
+            const error = new Errors.NoNextPage()
+            this.error(error.message, { exit: error.code })
         }
 
         await fcApi.apiCall("GET", {}, normalResponse)