+++ src/commands/recordings/download.ts
@@ -6,10 +6,10 @@ import { FreeClimbApi, FreeClimbResponse } from "../../freeclimb"
 import * as Errors from "../../errors"
 
 export class recordingsDownload extends Command {
-    static description = ` Download a Recording. Authentication is required to download a Recording, as with any other request made to the REST API. Returns a binary WAV audio file as an attachment in the HTTP response with mime-type audio/x-wav.`
+    static description = ` Download a Recording. Authentication is required to download a Recording, as with any other request made to the REST API.`
 
     static flags = {
-        next: flags.boolean({ char: "n", description: "Displays the next page of output." }),
+        next: flags.boolean({ hidden: true }),
         help: flags.help({ char: "h" }),
     }
 
@@ -45,25 +45,9 @@ export class recordingsDownload extends Command {
                 throw new Errors.UndefinedResponseError()
             }
         }
-        const nextResponse = (response: FreeClimbResponse) => {
-            if (response.data) {
-                out.out(JSON.stringify(response.data, null, 2))
-            } else {
-                throw new Errors.UndefinedResponseError()
-            }
-            if (out.next === null) {
-                out.out("== You are on the last page of output. ==")
-            }
-        }
-
         if (flags.next) {
-            if (out.next === undefined || out.next === "freeclimbUnnamedTest") {
-                const error = new Errors.NoNextPage()
-                this.error(error.message, { exit: error.code })
-            } else {
-                await fcApi.apiCall("GET", { params: { cursor: out.next } }, nextResponse)
-            }
-            return
+            const error = new Errors.NoNextPage()
+            this.error(error.message, { exit: error.code })
         }
 
         await fcApi.apiCall("GET", {}, normalResponse)