+++ test/commands/recordings-stream.test.ts
@@ -4,9 +4,6 @@ import { cred } from "../../src/credentials"
 
 const recordingId = "userInput-recordingId"
 
-import { data } from "../../src/commands/data"
-import { Environment } from "../../src/environment"
-
 describe("recordings:stream Data Test", function () {
     const testJson = {
         message: "Response from server",
@@ -119,96 +116,6 @@ describe("recordings:stream Data Test", function () {
             .command(["recordings:stream", "userInput-recordingId", "--next"])
             .exit(3)
             .it("Tests return of Exit Code 3 when flag next is not available")
-
-        const testJsonNext = {
-            total: 1,
-            start: 0,
-            end: 0,
-            page: 1,
-            numPages: 1,
-            pageSize: 100,
-            nextPageUri: null,
-        }
-
-        const nockServerResponseNext = `== You are on the last page of output. ==`
-
-        const testJsonNext2 = {
-            total: 2,
-            start: 0,
-            end: 0,
-            page: 1,
-            numPages: 2,
-            pageSize: 100,
-            nextPageUri: "",
-        }
-
-        const nockServerResponseNext2 = `== Currently on page 1. Run this command again with the -n flag to go to the next page. ==`
-
-        test.nock("https://www.freeclimb.com", async (api) =>
-            api
-                .get(`/apiserver/Accounts/${await cred.accountId}/Recordings/${recordingId}/Stream`)
-                .query({ cursor: "7265636f7264696e67733a73747265616d" })
-                .basicAuth({ user: await cred.accountId, pass: await cred.apiKey })
-                .reply(200, testJsonNext)
-        )
-            .stdout()
-            .env({ FREECLIMB_RECORDINGS_STREAM_NEXT: "7265636f7264696e67733a73747265616d" })
-            .command(["recordings:stream", "userInput-recordingId", "--next"])
-            .it(
-                "Test flag next works as expected when available with on last page",
-                async (ctx) => {
-                    expect(ctx.stdout).to.contain(nockServerResponseNext)
-                }
-            )
-
-        const finalCursor = "freeClimbCLIAutomatedTestCursor"
-
-        before(() => {
-            ;(async () => {
-                testJsonNext2.nextPageUri = `https://www.freeclimb.com/apiserver/Accounts/${await cred.accountId}/Recordings/${recordingId}/Stream?cursor=${finalCursor}`
-            })()
-        })
-        after(() => {
-            const dataDir = data.run([]).then(
-                (dataDir) => {
-                    const environment = new Environment(dataDir)
-                    environment.clearString("FREECLIMB_RECORDINGS_STREAM_NEXT")
-                },
-                (reason) => console.log(reason)
-            )
-        })
-        test.nock("https://www.freeclimb.com", async (api) =>
-            api
-                .get(`/apiserver/Accounts/${await cred.accountId}/Recordings/${recordingId}/Stream`)
-                .query({ cursor: "7265636f7264696e67733a73747265616d" })
-                .basicAuth({ user: await cred.accountId, pass: await cred.apiKey })
-                .reply(200, testJsonNext2)
-        )
-            .stdout()
-            .env({ FREECLIMB_RECORDINGS_STREAM_NEXT: "7265636f7264696e67733a73747265616d" })
-            .command(["recordings:stream", "userInput-recordingId", "--next"])
-            .it(
-                "Test flag next works as expected when available with has more pages",
-                async (ctx) => {
-                    expect(ctx.stdout).to.contain(nockServerResponseNext2)
-                }
-            )
-
-        test.nock("https://www.freeclimb.com", async (api) =>
-            api
-                .get(`/apiserver/Accounts/${await cred.accountId}/Recordings/${recordingId}/Stream`)
-                .query({ cursor: "7265636f7264696e67733a73747265616d" })
-                .basicAuth({ user: await cred.accountId, pass: await cred.apiKey })
-                .reply(200, undefined)
-        )
-            .stdout()
-            .env({ FREECLIMB_RECORDINGS_STREAM_NEXT: "7265636f7264696e67733a73747265616d" })
-            .command(["recordings:stream", "userInput-recordingId", "--next"])
-            .exit(3)
-            .it(
-                "Test error is caught when when using next flag and no reponse is given",
-                async (ctx) => {}
-            )
     })
 })