+++ CHANGELOG.md
@@ -9,6 +9,23 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 None
 
+<a name="0.3.0"></a>
+
+## [0.3.0] - 2021-04-07
+
+### Added
+
+-   callConnectUrl flag to command calls:make
+
+### Changed
+
+-   Descriptions of commands as well as flags and arguments
+
+### Fixed
+
+-   Incorrect instances of --next
+-   Incorrect instances of --acountId as query parameters
+
 <a name="0.2.2"></a>
 
 ## [0.2.2] - 2021-04-14